Last entry of the journey.
We finally added a like system, we created 3 account types, we added requests and reports, we added a sorting system for the species, we added identification capability to verifier users, we added moderator options for moderators.
We also created a video that explains in detail the usage of the app.
The report was also updated with all the info.